backend/requirements.txt
streamlit==1.32.0
firebase-admin==6.7.0
google-auth==2.38.0
langchain==0.3.23
langchain-community==0.3.21
langchain-google-community==2.0.7
langchain-google-firestore==0.5.0
langchain-ollama
langchain-qdrant
qdrant-client==1.14.2
python-dotenv==1.0.0
